backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.27.2
pydantic==2.5.0
chromadb==0.4.18
pandas==2.0.3
numpy==1.24.3
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.12.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
email-validator>=2.0.0
bcrypt==4.0.1
